consolidating converters (need to remove redundant / duplicate ones)
